* 0.11.x: (615 commits)
  CHANGELOG.md: Update for gh-3396
  DOC: log_progress: Reword cross-reference from 96b45b4bd
  ENH: adjust code comments
  ENH: adjust common_cfg description for datalad.ui.progressbar
  RF: use custom _naturalfloat for the best presentation of numbers
  ENH(DOC): cross reference LogProgressBar and log_progress in docstrings
  DOC: log_progress: Correct parameter name in docstring
  RF: versioned depends on sphinx within setup.py itself
  BF: do not bother estimating /sec if 0 sec passed
  RF/BF: reorder ProgressBarBase kwargs to match the one of the child
  BF(TST): making tests account for the changes
  NF: QuietConsoleLog to be used during tests (instead of SilentConsoleLog with no progress whatsoever)
  RF: AnnexRepo report to progress bar(s) that finish is partial if interrupted
  ENH+RF: LogProgressBar to log upon .finish()
  BF: report that gitrepo clone progress is about objects (not default to tqdm progress bar Bytes)
  ENH: datalad.ui.progressbar configuration
  BF: apply OP.MASK for git clone operation before considering
  CHANGELOG.md: Touch up 0.11.5 entries
  CHANGELOG.md: Second batch for 0.11.5
  RF: annexrepo: Tighten decision to commit with temporary index
  ...

+++ datalad/support/tests/test_path.py
@@ -7,11 +7,14 @@
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
 
+import os
 from ..path import (
     abspath,
     curdir,
     robust_abspath,
+    split_ext,
 )
+from ...dochelpers import exc_str
 from ...utils import (
     chpwd,
     rmtree,
@@ -20,6 +23,7 @@ from ...tests.utils import (
     assert_raises,
     eq_,
     with_tempfile,
+    SkipTest,
 )
 
 
@@ -27,6 +31,28 @@ from ...tests.utils import (
 def test_robust_abspath(tdir):
     with chpwd(tdir):
         eq_(robust_abspath(curdir), tdir)
-        rmtree(tdir)
+        try:
+            if os.environ.get('DATALAD_ASSERT_NO_OPEN_FILES'):
+                raise Exception("cannot test under such pressure")
+            rmtree(tdir)
+        except Exception as exc:
+            # probably windows or above exception
+            raise SkipTest(
+                "Cannot test in current environment: %s" % exc_str(exc))
+
         assert_raises(OSError, abspath, curdir)
         eq_(robust_abspath(curdir), tdir)
+
+
+def test_split_ext():
+    eq_(split_ext("file"), ("file", ""))
+
+    eq_(split_ext("file.py"), ("file", ".py"))
+    eq_(split_ext("file.tar.gz"), ("file", ".tar.gz"))
+    eq_(split_ext("file.toolong.gz"), ("file.toolong", ".gz"))
+
+    eq_(split_ext("file.a.b.c.d"), ("file", ".a.b.c.d"))
+    eq_(split_ext("file.a.b.cccc.d"), ("file", ".a.b.cccc.d"))
+    eq_(split_ext("file.a.b.ccccc.d"), ("file.a.b.ccccc", ".d"))
+
+    eq_(split_ext("file.a.b..c"), ("file", ".a.b..c"))
